+++ src/app.module.ts
@@ -1,7 +1,10 @@
 import { Module } from '@nestjs/common';
+import { MoviesModule } from './movies/movies.module';
+import { AppController } from './app.controller';
+import { App } from './app';
 @Module({
-  imports: [],
-  controllers: [],
-  providers: [],
+  imports: [MoviesModule],
+  controllers: [AppController],
+  providers: [App],
 })
 export class AppModule {}