+++ src/movies/dto/update-movie.dto.ts
@@ -1,0 +1,6 @@
+import { PartialType } from "@nestjs/mapped-types";
+import { IsString, IsNumber, IsOptional } from "class-validator";
+import { CreateMovieDTO } from "./create-movie.dto";
+
+export class UpdateMovieDTO extends PartialType(CreateMovieDTO) {
+}